Norm のエラーが出ないように parse.c を分割。
